Change RoPE implementation
This patch changes implementation of the RoPE, so that it is the
same as implementation in HuggingFace. Now instead of permuting
attention heads during the load of weights using HuggingFace
serializer, the library permutes attention heads only for a
reference Llama implementation.
