+++ kernel/rope.metal
@@ -38,23 +38,23 @@ rope(
     tensor2<const float> f_cos(params.freqs_cos_layout, params.freqs_cos);
     tensor2<const float> f_sin(params.freqs_sin_layout, params.freqs_sin);
 
-    const uint row_size = params.output.size(0);
+    const uint row_size = params.input.size(0);
     const uint head_dim = f_cos.size(1);
     const uint i = gid.y * threadgroup_size.y + tid.y;
     const uint k = gid.x * threadgroup_size.x + tid.x;
 
-    // numel = bs * seq_len * n_head * head_dim
+    // inputs size: (bs * seq_len * n_head, head_dim)
     const uint pos = i / (params.batch_size * params.n_head);
 
     if (i < row_size && k < head_dim) {
-        float x1 = params.input.at(i, 2 * k);
-        float x2 = params.input.at(i, 2 * k + 1);
+        float x1 = params.input.at(i, k);
+        float x2 = params.input.at(i, head_dim + k);
 
         float fcos = f_cos.at(params.start_pos + pos, k);
         float fsin = f_sin.at(params.start_pos + pos, k);
 
-        params.output.at(i, 2 * k) = T(fcos * x1 - fsin * x2);
-        params.output.at(i, 2 * k + 1) = T(fsin * x1 + fcos * x2);
+        params.output.at(i, k) = T(fcos * x1 - fsin * x2);
+        params.output.at(i, head_dim + k) = T(fsin * x1 + fcos * x2);
     }
 }
 
@@ -89,12 +89,12 @@ rope_freqs(
     const uint i = gid.y * threadgroup_size.y + tid.y;
     const uint j = gid.x * threadgroup_size.x + tid.x;
 
-    if (j < params.dim / 2) {
-        T freq = T(1.0) / metal::pow(params.theta, 2.0 * j / params.dim);
-        T angle = T(params.start_pos + i) * freq;
+    if (i < f_cos.size(0) && j < params.dim / 2) {
+        float freq = 1.0f / metal::precise::pow(params.theta, 2.0f * j / params.dim);
+        float angle = float(params.start_pos + i) * freq;
 
-        f_cos.at(i, j) = metal::cos(angle);
-        f_sin.at(i, j) = metal::sin(angle);
+        f_cos.at(i, j) = T(metal::precise::cos(angle));
+        f_sin.at(i, j) = T(metal::precise::sin(angle));
     }
 }
